requirements.txt
Django==3.0.6
django-leaflet==0.26.0
python-social-auth==0.3.6
django-oauth-toolkit==1.3.2
djangorestframework==3.11.0
drf-yasg
django-cors-headers==3.2.1
# -e git+https://github.com/autogestion/django-multiselectfield.git#egg=django-multiselectfield
django-multiselectfield
python-memcached==1.59
# psycopg2==2.7.4
psycopg2-binary
geoip2==3.0.0
#locustio==0.7.4
#-e git+https://github.com/pmdarrow/locust.git#egg=locustio
gunicorn==20.0.4